FAQ-Remove subsections for the git stuff
With the javascript version you don't actually get any information
before clicking on the link for the subheading... and that's not
very user friendly.

For now we'll just stuff it all in the single FAQ entry.  If it's too big
maybe we should split out the git stuff into it's own tutorial -- but that's
not really related to org-mode specifically... and the topic is huge and
probably covered better elsewhere on the net.