Added on FAQ on how best to organize org-mode files.
This question is one of the first questions that new users of org-mode face and
has come up from time to time on the mailing list.

This FAQ is meant to provide ideas to new users of org-mode.

It stresses the flexibility of org-mode's organizational structure and the power of agenda commands.