Add a link on the FAQ that goes back to the index page for the site
When visiting the FAQ I want to go to other pages on the worg site
and end up manually editing the URL to get there.  All of the
other pages seem to have a link back to the index page at the top
so I'm adding this back in to make it more consistent.