Do not use the ID property frivolously
Replaced all usage of ID with either task_id or resource_id as ID has
a different meaning and implies globaly unique ids which are found
across files. This is certainly not what is done with task_ids and
resource_ids which are only searched for in their respective trees
(actually resource_ids are not even looked for, it is left to
taskjuggler to match them).