ical2org: better timezone parsing
date and time information can come in different formats:
  date + T + time      floating (=local) time
  date + T + time + Z  always in UTC time

see section 4.3.5 of the ical spec http://www.ietf.org/rfc/rfc2445.txt

The third option is adding a TZID that specifies the timezone and a date+T+time in that time zone. This can't be parsed correctly at the moment, i.e. it is assumed at the moment that the TZID is the local time zone and no offset is applied.