org-hacks.org: Reorganization, added "copy link dest to killring" hack
Rationale:
* Remove "Enhancing the Org experience"
Everything in org-hacks enhances the org experience.
This headline conveys no information, even worse, it
can easily end up a kitchen sink for all kinds of
information that eventually no one will find anymore.
Distributed hacks to new headlines:

* New Headlines
- Structure Movement and Editing
- Org Table
- Capture and Remember
- Handling Links

* Remove Speed Commands
Speed commands are no hack but a regular feature for quite
a while now. So far, there was only a link to the manual
and "add your stuff here". Definately worth a headline
when there is actually content to put in there.