Move the "mark item done when all checkboxes checked" from org-faq to org-hacks.
Plant a link in org-faq to get to the hack.
Change the add-to-list to add-hook.
Wrap the add-hook within an (eval-after-load 'org-list ...) to make
sure that the hook variable is defined when we go to use it.
Also add a link to the documentation re. "statistics cookies".

Thanks to Peter Salazar and John Hendy for pointing out the defects.