Fix the example in the Export to XHTML section
With the new exporter, I was getting a complaint from
hack-local-variables (called from set-auto-mode in
org-html-final-function).

Nicolas pointed out that there should be a space after each colon in
the example. That does indeed make the example a <pre class="example">
... </pre> block. hack-local-variables still processes it as a local
variables section, which is arguably wrong, but it does not complain
any longer.