org-syntax: Beautify document
* dev/org-syntax.org: Remove trailing whitespace where it does not
change meaning.
(Headlines and Sections): There is one place where a source block is
used for an example of an Org document.  Change into an example block.
(Affiliated Keywords): There is one place in the file where lower case
for the delimiters of an example block is used.  Change into upper
case.