languages.org: Merge language support table, move language doc links
Most people use the table by searching for a particular language.  It
is easier to search in one sorted table than three.  This merges the
table so that the items are sorted continuously from top to bottom,
and adds a column to indicate the distribution method (which used to
be indicated by which section the language was in).

Also move the documentation links onto the language name instead of
having a separate column for it, and give language identifiers
verbatim markup.