index.org, babel/index.org: Moved link to Library of Babel
Library of Babel was referenced at the same place as the first
reference to Babel itself, but should really be something that viewers
discover after being intruduced to Babel.  This moves the link into
the Babel page.  This also rewords the top level description of Babel
to mention source code blocks.