org-contrib/*.org: Remove HTML_HEAD attribute to fix css
Many files were setting the HTML_HEAD attribute which was preventing
them from picking up the default css.  Perhapse the HTML_HEAD
attribute was once additive but now it replaces the configured value
and HTML_HEAD_EXTRA is used to add to it.  The div they were trying to
style doesn't exit so it would have had no effect anyway, so this
removes it, allowing the default configured HTML_HEAD to pull in the
default worg css.