*.org, worg.css: Fix or add HOME and UP links
The org headers contained LINK_HOME and LINK_UP attributes, but they
weren't included in the exported HTML.  The current HTML exporter
looks for HTML_LINK_HOME and HTML_LINK_UP.  This renames them on pages
with the old names, and adds them to some pages where they were
missing.

After that the links were added to the page but were hidden behind the
TOC, so the CSS was modified to nudge the TOC down a bit.