worg.css: Set body#content max-width to 48em
The original plan was to use 60em, but in a half-1080p-width browser
window, that causes the text to overlap with the unicorn image in the
lower-right corner of the page.  If that image is ever moved, this
width could be increased.

See <https://lists.gnu.org/archive/html/emacs-orgmode/2021-09/msg00403.html>.