org-maintenance.org: Add a note about bugfix releases
It's not easy to know what kind of bugfix should go in a bugfix
release, compare to those who should go in a minor release.

Bugfix releases only contain critical *and* trivial bugfixes.

This is because the person in charge of sync'ing Org-mode latest
release with Emacs code needs to know that it is always okay to
sinc a bugfix release within Emacs code.

When Emacs comes close to a release, only critical and trivial
changes are accepted.