org-faq.org: Recommend cb_thunderlink Thunderbird add-on
* org-faq.org (Thunderbird emails): Add CUSTOM_ID property to have
stable anchor during HTML export.  Recommend cb_thunderlink extension as a
means to generate mid: links and to handle existing thunderlink: ones.
Add a note that on Windows Thunderbird should register itself as a
handler of the mid: protocol.

Thanks to Bruno Barbier for discussion on the emacs-orgmode
mailing list.