* org-syntax.org (Text Markup): Add backslash to list of POST characters
Backslash appears to be supported. To test I used the following test
document:
```
foo ~bar~\& baz
```
The ampersand was chosen since that is not a supported POST character,
to make sure backslash was not simply escaping the next character.

In the documentation I wrote out the word "backslash" in parenthesis
to disambiguate between backslash and escaping the following comma.