org-contribute.org: Be a bit more welcoming to newcomers
* org-contribute.org: Explicitly state that we welcome all skill
levels to contribute and allow mistakes.  The idea is to address the
common fears of new free software contributors.  Move away the warning
about maintenance burden.  Only mention it for actual feature
contributions to not discourage the new contributors.
(As a contributor, what can I expect?): Move the maintenance
disclaimer here.